requirements.txt
python-dotenv==1.0.0
openai==1.6.1
langchain==0.0.354
asyncio==3.4.3